review feedbacks
* added shared.types.ts
* moved sitemap types to shared types
* revert fetchData refactoring
